refactor(team-provisioning): extract runtime liveness and identity
